+++ tests/wikidata/test_extract.py
@@ -17,13 +17,19 @@ from tests.wikidata.conftest import TESTDATA_DIR
 
 @pytest.mark.integration
 def test_search_organization_by_label() -> None:
-    """Test organization search in wikidata."""
-    expected = "Q26678"
+    expected = "Q679041"
 
-    search_results = list(search_organization_by_label(item_label="BMW"))
+    search_result = search_organization_by_label(item_label='Robert Koch Institute"')
 
-    assert len(search_results) == 2
-    assert search_results[0].identifier == expected
+    assert search_result, f"No organizations were found for id: {expected}"
+    assert search_result.identifier == expected
+
+
+@pytest.mark.integration
+def test_search_organization_by_label_for_none() -> None:
+    """Test if None is returned when multiple organizations are found."""
+    search_result = search_organization_by_label(item_label="BMW")
+    assert search_result is None
 
 
 @pytest.mark.usefixtures("mocked_session_wikidata_query_service")
@@ -32,9 +38,6 @@ def test_search_organization_by_label_mocked_error(monkeypatch: MonkeyPatch) -> 
     expected_query_response = [
         {
             "item": {"foo": "uri", "bar": "http://www.wikidata.org/entity/Q26678"},
-        },
-        {
-            "item": {"foo": "uri", "bar": "http://www.wikidata.org/entity/Q821937"},
         },
     ]
 
@@ -48,14 +51,13 @@ def test_search_organization_by_label_mocked_error(monkeypatch: MonkeyPatch) -> 
     )
 
     with pytest.raises(MExError):
-        _ = list(search_organization_by_label(item_label="BMW"))
+        _ = search_organization_by_label(item_label="BMW")
 
 
 @pytest.mark.usefixtures(
     "mocked_session_wikidata_query_service", "mocked_session_wikidata_api"
 )
 def test_search_organization_by_label_mocked(monkeypatch: MonkeyPatch) -> None:
-    """Test(mock) organization search in wikidata."""
     expected_item_details_response = {
         "identifier": "Q26678",
         "labels": {
@@ -160,6 +162,42 @@ def test_search_organization_by_label_mocked(monkeypatch: MonkeyPatch) -> None:
         {
             "item": {"type": "uri", "value": "http://www.wikidata.org/entity/Q26678"},
         },
+    ]
+
+    def mocked_query_response():
+        return expected_query_response
+
+    monkeypatch.setattr(
+        WikidataQueryServiceConnector,
+        "get_data_by_query",
+        lambda self, _: mocked_query_response(),
+    )
+
+    def mocked_item_details_response():
+        with open(TESTDATA_DIR / "items_details.json", encoding="utf-8") as f:
+            data = json.load(f)
+            return data[0]
+
+    monkeypatch.setattr(
+        WikidataAPIConnector,
+        "get_wikidata_item_details_by_id",
+        lambda self, _: mocked_item_details_response(),
+    )
+
+    search_result = search_organization_by_label(item_label="TEST")
+
+    assert search_result
+    assert search_result.model_dump() == expected_item_details_response
+
+
+@pytest.mark.usefixtures(
+    "mocked_session_wikidata_query_service", "mocked_session_wikidata_api"
+)
+def test_search_organization_by_label_for_none_mocked(monkeypatch: MonkeyPatch) -> None:
+    expected_query_response = [
+        {
+            "item": {"type": "uri", "value": "http://www.wikidata.org/entity/Q26678"},
+        },
         {
             "item": {"type": "uri", "value": "http://www.wikidata.org/entity/Q821937"},
         },
@@ -185,15 +223,13 @@ def test_search_organization_by_label_mocked(monkeypatch: MonkeyPatch) -> None:
         lambda self, _: mocked_item_details_response(),
     )
 
-    search_results = list(search_organization_by_label(item_label="TEST"))
+    search_result = search_organization_by_label(item_label="TEST")
 
-    assert len(search_results) == 2
-    assert search_results[0].model_dump() == expected_item_details_response
+    assert search_result is None
 
 
 @pytest.mark.integration
 def test_get_organization_details() -> None:
-    """Test organization details fetching from wikidata."""
     expected = {
         "identifier": "Q679041",
         "labels": {
@@ -294,7 +330,6 @@ def test_get_organization_details() -> None:
     "mocked_session_wikidata_query_service", "mocked_session_wikidata_api"
 )
 def test_get_organization_details_mocked(monkeypatch: MonkeyPatch) -> None:
-    """Test(mock) organization details fetching from wikidata."""
     expected = {
         "identifier": "Q26678",
         "labels": {